Switched to using official Clowder image.
Also removed the volume mapping. It works fine without it, because there
was not anything stored in it yet, and metadata is all in Mongo. Once we
want to use multiple instances of Clowder and things go into storage
we'll have to test again, and maybe try a remote volume driver (like
Cinder).